\name{survplot}
\alias{survplot}
\title{ggplot for survival data}
\usage{
  survplot(ss, covar)
}
\arguments{
  \item{ss}{a Surv object with right-censoring}
}
\value{
  a ggplot object
}
\description{
  plot survival data using ggplot
}
\details{
  lets use ggplot for this
}
\author{
  Barry S Rowlingson
}

